[mail/write] refactor selection of from tag
In the keyring it should possible to add a personal name for the account.
This is not visible in the UI, yet, but it could be set by OX6. ATM it’s
also not possible to choose another "identity" (i.e. mail account/alias)
to be used in the "From" field. The default is to use the account, which
contains the current or a default folder.

The new code is somewhat more readable, since it uses objects instead of
arrays and relies on data- attributes of the selected option element.

FIX: #24073