repopulate the add button in portal settings
after changing the widget list shown on the portal, the button menu to
add new widgets isn’t refreshed. This causes some menu items to be
invalid, when widgets have been removed or added to the portal. This
patch activates a refresh of the button menu, so only valid items are
shown.

FIX: #24503