make the tweet button more nice
No url should be shared (default was HTTP referer, which was annoying!),
but there is also i18n available, so we should use it.

FIX: #24562

The Bug report asked about intention of showing a twitter page instead
of providing an own UI for tweeting. AFAIK the twitter API doesn’t allow
this. It’s possible to provide some text for the tweet, but in the end
the user is still redirected to the twitter page to actually send the
tweet.