only show aliases for default account in options
The default sender that is configured here, is only valid for the
default account. It isn’t usable for any other accounts. This might be a
little strange to the user, because most users won’t see any other
e-mail than their default. This is due to a limitation in the design of
accounts on the backend.

bug:fix #24560