[possible unsafe] refactor carousel
This brings a more speakable name for internal method and slightly
change its behaviour. If the file object contains an URL field, this one
is used, instead of a call to the files API to get the URL. This allows
external users of this feature to provide image URLs itself.