provide api of modules selected by user via baton
After the import, the API needs to invalidate caches, so new items are
recognized. So, when creating a new format, you can now provide the API,
so events can be triggered. If you don’t provide the API object or the
object doesn’t behave like a standard API, a global refresh will be
triggered.