set sender based on selected folder when composing
there has been a timing problem, when composing a new mail. The form has
been rendered after the mail was set in the application. So the sender
fields haven’t been rendered, when the from should have been set. So
when composing a new mail, sender neets to be set again, when UI is
loaded completely.

bug:fix 25060