set sender based on selected folder when editing
there has been a timing problem, when editing a draft mail. The form has
been rendered after the mail was set in the application. So the sender
fields haven’t been rendered, when the from should have been set. So
when editing a draft mail, sender neets to be set again, when UI is
loaded completely.