change behaviour of filter function for folder
Never return any empty collection, if the user has some publications or
subscriptions. There are a few rare cases, when there is a folder marked
as "has publication" but in fact there is none. In these cases, users
might be confused if they click the "see publications" button and the
lists are empty. So we want to present all instead.