switch to baton in pubsub/settings/detail point
The data provided to the draw function is of type baton. So for
consistency we use baton. Also, the baton can contain a folder, so the
collections are filtered for this folder. If the result is empty, the
whole collection will be presented, so we should be fine to always call
the filter method.