be more strict about the info message
when dealing with filters in pubsub list, there are different cases. One
case is, that the folder might contain (global) subscriptions or
publications, but these are not owned by the user. An appropriate
message should be shown, if the user displays pubsub for a folder and
there aren’t any items found, that belong to him. In all other cases, it
might be the case, that the user just doesn’t have any publications or
subscriptions.