better check for pre-defined from address
a mail can have a from field set, but it could be an empty array. If
there is something set on the server, the from field contains an array
with exactly 2 elements. If this is not the case, the result from the
getPrimaryAddressFromFolder call should be used to set the from field.

Fixes a bug, that current folder is ignored for auto pre-selecting the
from field.

bug:fix #25310