fix pubsub settings pane
can’t say, what went wrong here or why this doesn’t work, but well. Now
it works, in the fail case, it might happen, that the item isn’t
rendered again, but the model will be updated and since the view will be
repainted on the next load, this won’t be a big deal.