reactivate persistant all cache
all and list caches have been set up to be non-persistant, because of a
huge performance decrease in some situations. Further analysis showed,
that only the list cache is problematic here. List cache has a huge
amount of data written in bursts after a request is finished. The queue
implementation if indexedDB could be better, so read operations are
preferred during a long write transaction. Since this isn’t the case,
every read during a burst-write takes very long.