only send password if it contains data
This will break with empty passwords, since its not possible to switch
from some password to an empty password. But since this use-case is
highly unlikely, this should be okay.

The reason to do this is, that we don’t have the password in the model
(for obvious reasons) and if we send null as the password, it will be
changed on the backend. This is most of the time not, what the user
intended.

bug:fix #25489