change logic to hide pop3 account settings
when creating a new mail account, no protocol is selected in the model,
yet. But imap is selected in the form, as a default. So the pop3 related
fields also need to be hidden, if the protocol is undefined in the
model. It is just more logical to check if the protocol is pop3 and hide
the pop3 related fields if this is not the case.

bug:fix #25447