use less aggressive mem caching
the model never refreshed during a session, so always the version from
memory was loaded. Now the model is synced with every access to the
model. This is not a problem, since the pubsub API takes care of the
caching. Since there might be changes on the server, it’s better to sync
with the server from time to time. The API takes care of this.