don’t leak event listener references
initialize is called multiple times, so every time a listener is
registered. This unregisters every time, since we don’t lack a new
version of backbone, we need to do it a little complicated.

See [1](http://blog.rjzaworski.com/2013/01/why-listento-in-backbone/)
why we would want to update backbone and use listenTo