add empty defaults for pubsub settings
no defaults are needed, yet, but sometimes it is possible, that some
module tries to load the defaults. It will then fail silently and the
module won’t be loaded. Our code should be more robust against this case
since commit 64debcda1a94b53f, but to be on the safe side, we just add
the file.

bug:fix #25695