fix wrong sender address when using inline links
data might contain the account_id, in this case, this should be
preferred. This is the case, when the compose dialog is created via
links that have associated a mail to them. This fixes wrong sender
addresses when using any of the inline links in a folder other than one
of the primary account.

This is a fixup commit for aea3acb0a4d3eb32f2b80e40aacd1284a9b8d77f