Fixed: Bug 26062 - Active App Suite frontend occupies a half CPU core without any actions
The problem was with the new "spinning" refresh icon. Having the spin
class and the spin-pause class at the same time caused the browser to
still calculate the animation. The fix removes the icon-spin class when
icon-spin-paused is active. Also the color shouldn’t change, since it
looks ugly when spinning and changing colors