don’t overwrite body when saving as draft
when saving as draft, the mail is fetched from the server because of
Bug 26086. This replaced the body text with the text from the saved
mail. This results in very annoying behaviour with the "auto save as
draft" feature, because the focus and the selection is lost. Since in
this case the text is not modified, we don’t overwrite it.