Fixed: Bug 26363 - Attachments are not removed from a draft.
Needed to refine the check to select the input field containing the
file. I’m not sure, where a field of type file with a value of
C:\fakepath/<filename> came from. The filter is needed to prevent
removing of the input field to add a new file. We only want to remove
input fields that actually contain a value.