added icon for windows 8 start screen added new ios icons for home screen added new ios splashscreens for fullscreen mode
Windows 8 needs a 144x144px icon which will be displayed on the tile
if the user pins App Suite to the start screen.

iOS needs icons for each device size (iPhone 3,5" and 4", retina and non-retina)
Also splashscreens must be provided in 3 sizes to match all phone sizes. For iPad we do not use any splashscreens as the files will get to large