Fixed: Bug 26383 - portal access still possible via URL even if disabled at the backend
when launching an application, check if the app is registered with the
manifestManager. This is not the case, if the required capabilities are
not met, so the app should never load, if it is not registered at the
manifestsManager.

Befor this fix it was possible to launch applications using the
autoLaunch feature (provide app=<application> in the url) even if the
required capabilities weren’t met.