Fixed: Bug 26364 - TSK-0019 Cannot find task 9 in context 17
Change: do not always trigger a local refresh event

the event should IMHO only be triggered when the action is local. If
there are also changes on the server, refresh.all should be triggered.
The problem was, that on refresh:all:local also a refresh.all event is
triggered. In most cases, this will trigger an all request to the
server. This is of course done _before_ the delete request and the
server will return a list containing the element that is to be deleted.

refresh.all will still be triggered _after_ the delete request is
finished, which is the correct place.