implemented US:swipe gesture is enabled in listviews to perform custom actions
VGrid now supports swipe left and swipe right on mobile devices.
The Grid now has two new options, the swipeLeftHandler and the
swipeRightHandler. Provide a function for these in your options that
 will be executed if a swipe is triggerd on a Vgrid cell.

This feature is used to show a delete button on the cell
in mail, contacts and tasks after the user has swiped the
cell in right direction.