remove tinyMCE ox skin background reset class
many elements in the ox skin had a background set to transparent. It
wasn’t possible to use the emoji icons, because this class had more
importance than the emoji css class.

This change is possible unsafe as it might have some side effects. I
tested as much as possible and everything seems to work fine.