Fixed: Bug 26296 - [L3] Attached .eml file disappears when repeat the save of the mail.
the fix is to handle the nested_msgs array also before saving as draft,
not only before sending the mail. This fix does a refactoring of cleanup
tasks that need to be done before the mail is send to the server. These
things need to be done before saving as draft and before sending the
mail and are now unified in a new method.