render collection controls only if needed
the tabs are only needed if the setting is provided by the backend and
there is support for softbank emojis.

the dropdown menu should only be shown if there is a list of collections
to choose from.
(cherry picked from commit b9e6aac77d22768b15cf82ff40467884bfae34d4)