preserve emoji unicode when switching mail format
text -> html worked fine before, but for html -> text emoji icons were
always dropped, because they have been part of the text as images.

I could not figure out, how to change the content of the selection,
before it was returned by editor.getContent({format: 'text'}). So the
solution is to get the html content and parse it using jQuery after
transforming all emoji img tags to corresponding unicode.