never append vcard when saving mail as draft
The backend will append a vcard every time a mail is saved. If a vcard
is already added as an attachment from the previous save operation, the
vcard is attached anyway, resulting in a growing list of attachments.

This fixes a part of
Bug 27156 - Auto-save during mail compose messes things up