emoji: return undefined instead of special object
This changes the API, but since the API is considered experimental, this
should not be a problem. This return value is not used anywhere, yet.
Specs to test this (new) behaviour will be added soon.

This change will prevent to create a new object, even if no information
for an icon can be found and will returned undefined instead. This seems
more consistent to me.