fix potential issue with settings in one example
settings.set({someting: 'some value'}) will overwrite the existing
setting. In this case it was intended to set both of the variables. The
example didn’t fail, because checks being not hard enough.

Anyway, doing it correctly is better, since other developers might
copy&paste code and doing it wrong.