emoji: add fallback for icons to other icon sets
If no icon is found in the current emoji collection, there is now a
fallback mechanism.

The first priority will be the currentCollection, then the available
collections will be searched one by one, in the order of occurence
within the string from left to right.

If the backend setting 'overrideUserCollection' is set to true, only the
available collections will be searched.

This allows you to override the precedence of the icon sets. If you want
to force an icon set onto a user, set 'overrideUserCollection' to true
and specify your preferred collection in 'availableCollections' first.