spec/vgrid: remove expectation about initial selection
this should not be in this example, since the initial selection is not
part of it. This can be added in another example, if needed.

Basically fixes this example in chrome, because it behaves slightly
different than phantomjs. May be due to a setting (last selected item is
rembered in the settings). But anyway, this would be another issue and
should be tested in another example.