Fixed: Bug 27588 - When clicking checkbox it disables
fixes a failing spec. The problem was with selectSmart() and multiselect
enabled, it was not possible to have an empty selection, because the
focus event is triggered often but not often enough. Since we want to
allow "no selection at all" when multiselect is set to true, we need to
disable selectSmart() in this case.