Partly Fixed: Bug 27642 - "Expand" indicator shown for folders without subfolders
this is a workaround for a backend bug, reporting a folder has children,
but there are none. In this case the arrow indicating subfolders is
hidden, when no children have been loaded. But if no children have been
loaded, yet, the arrow might still appear until the request has been
made.