Fixed: Bug 26383 - portal access still possible via URL even if disabled at the backend
Add a new hash object to track disabled apps. Apps can be disabled by
adding the name as a key and true as the value. A value of 'upsell'
might be used to indicate upsell capabilities for this app but it’s not
used, yet.
(cherry picked from commit ef4bad2b2c1991ba0b2a81c686660d50bc82ec68)