refactor IndexedDB based cache implementation
This is quite a major change and might need some review.

The previous implementation created new databases for each cache and
those were managed using a meta-database collecting all the databases in
order to track all existing caches. This was needed, because there is no
way to find out about existing databases using the IndexedDB API. This
implementation lead to various problems when cleaning up all the
databases due to the asynchronuous API of IndexedDB.

The new implementation uses only one database and adds prefixes to the
keys to identify the cache used. Specs are still passing and the
implementation seems to work fine. I was not able to verify that
everything is cached correctly or if some data is not cached because of
some bugs in the code. Of course, this code is completely bug free and
working as expected.