Fixed: Bug 27896 - [L3] Error shown after login on mobile devices
There is now a new setting 'autoStartMobile', that can contain an app
that is started on small devices. For larger devices 'autoStart' setting
will be used.

Another thing that is different on small devices now is that from the
autoLaunch array (that can contain many apps that are automatically
loaded during startup) only the first app loadable app will be started.