Fixed: resizing of emoji icons possible in ff + ie
Emoji icons should not be resizable. Fixed by disabling object resizing
for the editor. This is the most easy way to fix it for tinyMCE 3.4.x -
in 4.x you would set the data-mce-resize attribute to false, to disable
this feature for one item. Once we switch to tinyMCE 4.x,
object_resizing: 0 can be removed from tinyMCE settings and everything
will still work fine.