spec/emoji: add failing conversion example
This is a bug when doing a conversion from imageTags to PUA encoding.
The converted text might not be html code, but plain text. Converting a
">" would then encode this character (and others) with html entities.
This should not happen. Same for HTML code that is not an image-tag with
emoji information, it should not change.