Fixed: convert from unified to PUA with special chars
When converting plain text messages from unified to PUA encoding and the
plain text has some special chars like "<" or ">" in the text, the
parsing will break the plain text. This is, because jQuery has been used
for parsing, which is not a good idea for non-html texts. Now, parsing
is done without jQuery.

Specs from develop branch are still passing, but this code needs a
little refactoring. Anyway, this will be done in develop branch, only.