added new config param to force desktop-style app-launchers on mobile
"forceDesktopLaunchers" is an new core setting (io.ox/core) which will overwrite the standard behavior of app launchers on mobile devices. By default all launchers will be hidden in the laucher menu. On desktop devices the default is to use all available space of the upper toolbar and fill it with launchers. This can now also be enabled on mobile devices. (This might change nothing compared to the default if the first app title string is too long to display in the topbar.)