Fixed: Bug 28446 - jslob?action=set fired before mail requests
Optimise Settings::save to use update action, if values have only changed.
Fallback to set action, if properties have been removed.

This saves a lot of bandwith, since every change in a grid triggered a
complete save of the settings for a module. Now only the grid options
are send to the server.