hotfix for wrong processEmoji method
this fixes an issue, that 'undefined' is shown as the content of mails.
It’s a quite ugly hack and it won’t work for the first few calls of
process emoji (unprocessed mail is shown instead, which is better than
no content at all). But this would need some cleanup.