Fixed: Bug 28496 - [L3] emoji disappear when adding signature with IE10
Rely even less on jQuery parsing during transformation of emoji items.
The order of the attributes for a html node can not be calculated
deterministic on all browsers. It worked in IE9, but not IE10. Now it
should work in all browsers, since we are only using basic javascript
string operations, now.