Fixed: Bug 28427 - [L3] Mails in thread view are shown like unread in case 'Mark all mails as read' feature is used
This is still not a really good solution for this problem. But without a
larger change in a core class, it won’t be possible to fix it in a clean
way. I added a comment in the code, to document this.

Technically, the VGrid class needs a new method to indicate it’s loading
state. It should be a deferred object in pending state, if it is loading
and resolve with the grid object once loading is finished.