Fixed: Bug 28108 - no login possible if IMAP not accessible
Better error handling when setting default folder. Sometimes the default
folder might be null, in these cases, we inform the user that he should
expect problems, because the default folder could not be loaded. Anyway,
the app can be started just fine.

This happened, when the IMAP server was unreachable and the backend
didn’t send default folders for mail application. External E-Mail
accounts worked well in this case, so we don’t disable the application
but just continue loading.