US: don’t show hidden folders per default
This is a combination of 3 commits.

Folders starting with a dot should be hidden by default. It is also
possible to disable the filter with a user defined setting.

---

add option "show hidden files" to settings

new feature to filter some hidden files should be configurable by user.
So an option is added to the files settings to show hidden objects
(actually disable the filter).

---

refresh folder list on "showHidden" files option

the folder list won’t be up-to-date after the option has been changed,
until the refresh button is clicked. This automatically triggers a
refresh for all folders.