Fixed: Bug 29423 - [L3] emojis will be send in unicode6 on iOS
on emoji enabled devices, the convert function from unified to pua did
not work. There is now a test case to check exactly for this behaviour.
It fails without the patch to fix the bug and the bug is not
reproducable any longer with this patch applied.

The problem was, that the conversion method relied on unifiedToImageTag
method, which does nothing on emoji enabled devices. Conversion is now
forced, so there will always be image tags for all emoji.