Fixed: Bug 29363 - Contact Restore Capability
Don’t fail if method is not defined, yet. This callback will be called,
if the app is restored after a reload of the page using the failRestore
mechanism. During this initialisation, the getDirtyStatus method will
not have been defined, until initialisation is done. Anyway, this can
safely be ignored in this case.