Made the package parameter optional; misc. clean-up.
Without a package parameter and without debian/changelog, the generated
manifest is called manifest.json. To get rid of it when init-packaging
is executed later, init-packaging now runs the clean task.

Added a list of known licenses to init-packaging.
debian/rules is now made executable.
Lowercased expat.txt so that it can be found by init-packaging.