add deprecation warning to io.ox/mail/utils//getSmartTime method
This method breaks a test and seems to be unused, because it brings a lot of
problems to manually update the string without the page being reloaded. This
might confuse the user and therefore we decided not
to use this method any longer. It has not been removed,
yet but should so in the future.
The following warning is there to inform potential
3rd-party developers about the change.