fix specs for tasks API
same as for autocomplete API:

a change within the http API prevented a request from being sent, if the
payload has already been sent, before. In order to not have the request
cached, the description of the test is added to the payload, so it will
be different most of the time.