Fixed: Bug 29845 - [L3] data leaked via indexedDB between different users
Requests might be delivered from cache if 2 users with different
contexts share the same browser. In this case, id and folder_id
parameters don’t lead to unique requests any longer. Therefor a user
context needs to be added to the URL. The generated URLs are now unique
to one specific user in one specific context.