Fixed: Bug 30076 - [L3] Folder created while moving a task does not display on Move Task window
remove ‚add folder‘ action from all modal dialogs with folder list

hint: After talking to Matthias this seems to be the best solution. The create actions weren’t meant to occur in the modal dialogs and the equivalent mail/files dialogs don’t support this in general. Maybe there will be a follow up user story for this.