change behaviour of search widget in window class
this is a fixup for commit 7f6f544672fe25afc3455eb83978eb388b7d83d2,
since it didn’t really resolve the issue. After some research in the
version history ($.trim call has been there from the very beginning of
this feature), we decided to remove the call to $.trim. This will change
the default behaviour, but we were not able to think of any use-case,
when trailing whitespace might break a search.