Bug 29451 - [L3] Emoji is garbled at mail list view, however emoji-icon displays correctly when user comes back from another folder
io.ox/core/emoji/util::processEmoji had to be extended in order to fix
this bug. The API is backwards compatible, only a callback parameter is
added, that expects a callback function that will be called, once the
emoji library has been lazy loaded. It will also be called, if the
library has already been loaded, however, the callback function will be
notified about this in a second parameter (after the processed text).
This can than be used to implement asynchronuous behaviour of the emoji
processing but still use the synchronuous way, that has been used
before.