fix sending of emoji characters
this has been an undocumented bug, that broke sending emoji characters
inserted with our own emoji plugin for tinymce. The cleanup regexp
replaced important information on the img tags that was used to identify
emoji images and (more important) their unicode6 representation (stored
in data-emoji-uncode attribute of the img node.