fix folderAPI tests for some setups
this is a timing problem, on update, warn:hidden will be triggered. This is
not dependend on the deferred object that will resolve after the update is
finished. The only reliable way to test this, is to wait for the warn:hidden
event, so this should work for every setup, now.