add task to generate the ox.pot file
for the po template, all strings that should be translated are extracted from
the js sources and put into the pot file with some context of their
origin. This can then be used by translation tools to create the po
files that are compiled to require modules to be loaded by the
application and provide translations.