Fixed: Bug 30753 - Mail gets deleted on forward/reply
Completely remove any handling of auto saved mail, because this is
handled in the backend. This is a fixup for cec5c83595d5df37c042, which
causes possible data loss (the referenced bug!). The bug fixed by the
fixed up commit (Bug 30416 - draft mail is always deleted after sent)
should still be fixed, or is a backend bug.