Fixed: Bug 31014 - [L3] EmiG: Wrong sender address checked when editing a draft
During the 'open' event of the mail window, the mail data is not set
yet, so that only the default value of the select element is seen in the
"from" field.

The initial filling of the from field does not trigger any events
either.

Switched the initial sender check from immediate to an extension point
used for manipulators of mail data.