fix uglify dist task (mostly because of i18n)
there are i18n modules that should not be packaged. So the uglify tasks
need to be defined in grunt/tasks/dist.js, since in this file we know if
a language should be packaged or not.

All this still seems to complicated for me and it should be refactored.