Fixed: Bug 31040 - Build system does not compile dependent LESS files automatically
grunt-newer allows an override-method to determine if a file needs to be
handled even if it hasn't changed since the last run of the task.
This is the case for files depending on each other. For less files, there
has been an upstream solution to handle dependencies caused by @import
statements, this has been extended to handle our implicit dependencies,
too. (i.e. a themes' definition.less or mixins.less changes)