hide warnings during build if devDependencies are missing
we support to build and distribute everything without devDependencies installed.
Should be okay to have all optionalDependencies.

This uptade to latest release of shared-grunt-config also contains a bugfix in
less tasks, it should be save to install. Also brings a tiny new feature to
help generate grunt/local.conf.json