Fixed: Bug 32654 - [L3] After update OX App Suite UI can not be loaded - boot.js script does not respond
Root cause:
A bug in the timezone package on SLES.

Zoneinfo files in the package timezone-2014c-0.3.1 contain an additional
first transition at -0x80000000 seconds in the 32 bit version and at
-0x8000000000000000 seconds in the 64 bit version. Since especially
the transition in the 64 bit version is ca. 20 times older than the
universe, and the numbers are round binary numbers, this looks like a
bug in the generation of zoneinfo files. Also, the 2014c files on Debian
don't have this transition.

OX App Suite uses an array indexed by time to store the transitions. The
additional transition causes this array to grow larger than any
available RAM.

Solution:
Capped the size of the transition array at -1000 years.