fix issue with ignored "device" field in manifest files
if an app was in the favorites list, the device field specified in a
manifest.json was ignored. Fixed this by using the appropriate API
method to check if an app is disabled or not and for the case, where
this is explicitly ignored, added a manual check for the device field.