remove peerDependency for all third party plugins
when re-structuring our build and development system, we first wanted
to get rid of the dependency to a core build. This would have meant to
extract the default theme, so third party plugins could have depended
on this one. Since we dropped this goal for this release, the shared
grunt config now expects a coreDir option to be set, again. This means,
core must package the dependencies, so themes can build against
the core build.