Workaround: Bug 32764 - Privacy information is not translated
grunt-require-gettext can not extract dynamically created strings (even
if the result can be determined statically, as in this case). To get this
string into the pot file ASAP, using a "real" multiline string, as defined
by ES5 specification. However, this is considered unsafe by jshint, so this
is only a temporary solution until statically determinable strings are
supported by the extraction tool.