Fixed: Bug 32920 - Lots of debug output when logging in
Root cause:
Before 7.6, all assert() calls were completely removed for non-debug
builds.

Since we hope to get rid of debug builds with the help of source maps,
and the existing calls to assert() don't test anything complicated to
justify their complete removal, re-implementing this feature in Grunt
makes no sense.

Solution:
Switch to run-time check in the assert() function. Also, added on-demand
initialization of the translations in io.ox/core/strings.

If somebody wants to perform extensive checks for the assert() function,
extend it to recognize a callback function as first parameter.