Fixed: Bug 33287 - Deleted E-Mails keep coming back
This has been a side-effect of commit 3c989361f257e32158a37828e53092bc29
this is supposed to fix bug 33138 - thread count does not update after i
delete a mail within a thread.

I fixed this by reverting 3c989361f257e32158a37828e53092bc29 and added
code to trigger the delete event within the mail API, once the delete
request is done. There is already a handler registered to the delete
event internally. This takes care of the refresh.

One drawback is, that the delete request might take quite some time,
so the user is presented with an up-to-date UI, only after the delete
request finished. (Talking about multiple seconds, here.) Especially
for mails in a thread, if the first mail is deleted, this is not as
nice as it could be, since the UI looses the thread information during
that time. However, solving this issue seems very complicated to me,
could need quite large changes to the code, and therefor require
extensive testing.