fix gt call issue
only call gt with variable parameter in very special cases. In order to have
strings extracted from the sources, the parameter needs to be statically
evaluatable to a string.

Solution here: move the gt call to where the strings are defined, the
tokenfield extension now expects a translated string as first parameter.