fix 'isLarge' detection for fixAbsolutePositions method
the content variable did never contain any isLarge attribute, because it
contains a jQuery element. Seems this code used the baton earlier.
Simply provide the isLarge attribute as a second parameter for this
method. This one is internal, so it should not break anything elsewhere
in the code.

This is a performance optimization, that was developed for
Bug 33308 - big mails leads to script warning in firefox