optimise performance when rendering large mails
don't use jQuerys empty method, since it will generate huge amount of
memory for a short moment. This is a problem for very large mails, since
the browser will possibly run out of memory when performing the
operation. Using the native browser methods works fine, too and for
those browsers that don't implement getElementsByClassName, the jQuery
fallback will be used.

Note: possibly unsafe, needs review/extended testing. For a large mail
(4MB) I was able to reduce memory peak consumption from 1GB to ~50MB
during rendering. This will hopefully prevent firefox from showing
'script is not responding' popups and therefor fix
Bug 33308 - big mails leads to script warning in firefox

be more conservative during optimisations

commit a29ae4530475c62e45901adacf82a14cc9460577 introduced a quite
aggressive optimisation for mails with a large body. This fixup only
applies this optimisation on the actual content node (which is the large
part) and not the complete body. This way, handlers attached to the body
node will be cleaned up by jQuery in the empty method.

cherry-picked and squashed from
commit a29ae4530475c62e45901adacf82a14cc9460577 and
commit 0678fa44664f9cec47758c17da6cc5e431cba87a