choose default sender for new mail from current folder
when using external accounts, it's useful to set the address of this account
as default sender. This has been the behaviour defined in the old mail compose
application and should be used in the new one, too.

Code might need a little refactoring, though. Feels a little 'verbose' and
over-complicated. May be, a plain backbone view would also do the job with
bindings to the 'from' attribute of the model.