Fixed: Bug 34352 - Unread-toggle missing for mails on mobile devices
Beside the mark-unread action some other actions were missing in mail which some users might want to use on their mobiles, too. To be consistent to the actions used in the portal mail-sidepopup, a "more" button was introduced to the mail detail view's bottom toolbar. The 1-level actions are now reply, reply all, delete and forward. All other additional actions are listed under a dropdown in the toolbar. Also the unread state of a mail now get's shown with the desktop-style blue envelope next to the subject.