Fixed: Bug 34692 - No default Folder is displayed
after a behavioural change for Bug 34082, most folders have been reported
to be of type 'inbox'. Which is partly correct, because most folders are a
subfolder of the INBOX. So this commit reverts to the same behaviour for
type inbox, as before. For other types, the new behaviour is still
valid.

This does not seem to be a problem with the refactored folderview in the
latest develop version, but in version 7.6.0, the folderview seems to
make some assumptions about folders of type inbox.