Fixed: Bug 34774 - Folder tree is collapsed by default, new folders don't show up
Fixed it like in mail. Listening to an add event on the collection and then manually open the appropriate folder.
This fix also included to make small changes on how models are updated in the collection. If a reset on the collection
is required, this will no be triggered manually and add the models via set.