Fixed: Bug 34956 - Installation fails at "updating themes" due to OutOfMemory
in some cases, using the import statement in less files lead to
recursive loading of one and the same file over and over again. As a
result the search paths array grew longer and longer and eventually lead
to the OutOfMemory exception. Before adding a new path to the search
paths array, it should be ensured, that the path is not already part of
the search paths.