workaround: Bug 35439 - Inbox widget show: Mail folder "null/INBOX" could not be found on mail server
This has been introduced with the fix for Bug 35316
(f3fb4a96f136cf01b1f9521579a41dedcc6cf275). After the change in the backend to
support simplifying UI code returns a string "null". This was
interpreted as a valid name for a folder. This patch/workaround now
interprets "null" as an invalid or non-existing folder.