add optional parameters for update-themes to work more efficient
This adds two optional parameters to the update-themes script in order
to make the process more efficient. Calling update-themes.sh with the
--later option will be a very fast operation, now. It should be used
most of the time. The second option is --if-needed which will check for
a magic file and only run update themes if this file is present. Every
run of update-themes will remove this magic file, so running with
--if-needed will be fast, as long as the magic file is not present.

Running update-themes with no parameters at all will always update the
themes and remove the magic file (if present). However, this usage is
considered deprecated.

This implements chore US 76851366