Fixed: Bug 35599 - [L3] undescriptive Error message when sending mail and attachment was moved away or deleted
This fix is a workaround because we are not able to detect when a local file is deleted. This workaround informs the user if an empty file is uploaded when clicking send/save. Empty files are an indicator that this file could have been deleted.

(cherry picked from commit f3d33ec600b631f9d048cf8d9ffa0f68d6838093)

Conflicts:
	ui/apps/io.ox/mail/compose/view.js