Fixed: Bug 35952 - [L3] address containing <> in signature does not display part in <> if mail is set to text only
Use better filtering of signature code, which might be HTML or not. The
original Bug was reported against 7.6.1 code, which only had a plain text
area to edit the signature. With 7.6.2, there will be a tinyMCE based
editor for that. This fix works in both cases, if the signature code is
HTML based or not. It will strip out HTML tags for plain text mails
(trying to preserve line-breaks done with <p> and <br> tags) and encode
special HTML characters, like <, >, &, …

Porting this back to 7.6.1 might be a little tricky and would need
some testing. However, there are some tests, that will support any
efforts to backport this to 7.6.1.