fix problem with inline images when saving mails as draft
when manually saving mails as draft, the src attributes of inline images
are changed on the backend. Those need to be adjusted when the draft is
created on the backend, so they are not deleted automatically after some
timeout. However, the extraction did not work after a change some time
ago and is now fixed including some tests, that specify this behaviour.